Refactor pygame_gui into package
