+++ pygame_gui/view.py
@@ -1,1016 +1,62 @@
-"""Minimal Pygame interface for the Tiến Lên card game."""
-
 from __future__ import annotations
 
-from pathlib import Path
-from typing import Dict, Tuple, List, Callable, Optional
-from enum import Enum, auto
 import json
 import logging
-import math
-import types
+from pathlib import Path
+from typing import Dict, Tuple, List, Optional
 
 import pygame
 
-from tien_len_full import Game, Card, detect_combo, Player
+from tien_len_full import Game, detect_combo, Player
 import sound
 
-LOG_FILE = "tien_len_game.log"
+from .helpers import (
+    TABLE_THEMES,
+    PLAYER_COLORS,
+    HAND_SPACING,
+    HORIZONTAL_MARGIN,
+    LABEL_PAD,
+    BUTTON_HEIGHT,
+    ZONE_GUTTER,
+    AVATAR_DIR,
+    AVATAR_SIZE,
+    GameState,
+    calc_start_and_overlap,
+    calc_hand_layout,
+    list_music_tracks,
+    list_table_textures,
+    load_card_images,
+    get_card_image,
+    get_card_back,
+    CardSprite,
+    CardBackSprite,
+    draw_glow,
+    _mixer_ready,
+)
+from .overlays import (
+    Button,
+    Overlay,
+    MainMenuOverlay,
+    InGameMenuOverlay,
+    SettingsOverlay,
+    GameSettingsOverlay,
+    GraphicsOverlay,
+    AudioOverlay,
+    RulesOverlay,
+    HowToPlayOverlay,
+    TutorialOverlay,
+    SavePromptOverlay,
+    ProfileOverlay,
+    GameOverOverlay,
+)
+from .animations import AnimationMixin
 
 logger = logging.getLogger(__name__)
-if not logger.handlers:
-    handler = logging.FileHandler(LOG_FILE, encoding="utf-8")
-    formatter = logging.Formatter(
-        "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
-    )
-    handler.setFormatter(formatter)
-    logger.addHandler(handler)
-    logger.setLevel(logging.INFO)
-    logger.propagate = False
 
 
-def _mixer_ready() -> bool:
-    """Return True if pygame's mixer is initialised."""
-    return bool(pygame.mixer.get_init())
-
-
-TABLE_THEMES = {
-    "darkgreen": (0, 100, 0),
-    "saddlebrown": (139, 69, 19),
-    "navy": (0, 0, 128),
-    "darkred": (139, 0, 0),
-}
-
-# Color used for card glows by player index
-PLAYER_COLORS = [
-    (255, 255, 255),  # Human - white glow
-    (255, 100, 100),  # AI Left - red glow
-    (100, 100, 255),  # AI Top - blue glow
-    (100, 255, 100),  # AI Right - green glow
-]
-
-USER_DIR = Path.home() / ".tien_len"
-OPTIONS_FILE = USER_DIR / "options.json"
-SAVE_FILE = USER_DIR / "saved_game.json"
-
-# Default distance between cards in a player's hand
-HAND_SPACING = 20
-# Horizontal margin used when centring hands on screen
-HORIZONTAL_MARGIN = 20
-# Extra padding used when positioning player labels
-LABEL_PAD = 10
-# Button dimensions and layout spacing
-BUTTON_HEIGHT = 40
-ZONE_GUTTER = 10
-AVATAR_DIR = Path(__file__).with_name("assets") / "avatars"
-AVATAR_SIZE = 40
-
-# Helper for positioning card sequences
-
-
-def calc_start_and_overlap(
-    screen_width: int,
-    count: int,
-    card_width: int,
-    min_overlap: int,
-    max_overlap: int,
-) -> tuple[int, int]:
-    """Return `(start_x, overlap)` centred within `screen_width`.
-
-    `overlap` is clamped to the given range and represents how much consecutive
-    cards cover one another."""
-    if count <= 0:
-        return screen_width // 2 - card_width // 2, 0
-    min_spacing = card_width - max_overlap
-    max_spacing = card_width - min_overlap
-    if count > 1:
-        fit_spacing = (screen_width - card_width) / (count - 1)
-    else:
-        fit_spacing = card_width
-    spacing = max(min_spacing, min(fit_spacing, max_spacing))
-    total_w = card_width + (count - 1) * spacing
-    start_x = screen_width // 2 - int(total_w // 2)
-    overlap = card_width - spacing
-    return int(start_x), int(overlap)
-
-
-def calc_hand_layout(screen_width: int, card_width: int, count: int) -> tuple[int, int]:
-    """Return ``(start_x, spacing)`` for a horizontal hand."""
-
-    start_rel, overlap = calc_start_and_overlap(
-        screen_width - 2 * HORIZONTAL_MARGIN,
-        count,
-        card_width,
-        25,
-        card_width - 5,
-    )
-    spacing = card_width - overlap
-    return start_rel + HORIZONTAL_MARGIN, spacing
-
-
-# ---------------------------------------------------------------------------
-# Helpers for loading and caching card images
-# ---------------------------------------------------------------------------
-
-_CARD_CACHE: Dict[Tuple[str, int], pygame.Surface] = {}
-_BASE_IMAGES: Dict[str, pygame.Surface] = {}
-
-
-def list_card_back_colors() -> List[str]:
-    """Return available card back color names."""
-    backs_dir = Path(__file__).with_name("assets") / "card_backs"
-    colors: List[str] = []
-    for img in backs_dir.glob("*.png"):
-        stem = img.stem
-        if stem == "card_back":
-            colors.append("blue")
-        elif stem.startswith("card_back_"):
-            colors.append(stem.replace("card_back_", ""))
-    return sorted(colors)
-
-
-def list_table_textures() -> List[str]:
-    """Return available table texture names."""
-    tex_dir = Path(__file__).with_name("assets") / "tables"
-    return sorted(p.stem for p in tex_dir.glob("*.png"))
-
-
-def list_music_tracks() -> List[str]:
-    """Return available music track filenames."""
-    mdir = Path(__file__).with_name("assets") / "music"
-    return sorted(p.name for p in mdir.glob("*.mp3"))
-
-
-class GameState(Enum):
-    """Simple enum representing the game's current state."""
-
-    MENU = auto()
-    PLAYING = auto()
-    SETTINGS = auto()
-    GAME_OVER = auto()
-
-
-def _image_key(card: Card) -> str:
-    rank_map = {"J": "jack", "Q": "queen", "K": "king", "A": "ace"}
-    rank = rank_map.get(card.rank, card.rank.lower())
-    suit = card.suit.lower()
-    return f"{rank}_of_{suit}"
-
-
-def load_card_images(width: int = 80) -> None:
-    """Load all card images scaled to ``width`` pixels."""
-    assets = Path(__file__).with_name("assets")
-    for img in assets.glob("*_of_*.png"):
-        key = img.stem
-        base = pygame.image.load(str(img)).convert_alpha()
-        _BASE_IMAGES[key] = base
-    backs = assets / "card_backs"
-    for img in backs.glob("*.png"):
-        _BASE_IMAGES[img.stem] = pygame.image.load(str(img)).convert_alpha()
-    for key, base in _BASE_IMAGES.items():
-        ratio = width / base.get_width()
-        _CARD_CACHE[(key, width)] = pygame.transform.smoothscale(
-            base, (width, int(base.get_height() * ratio))
-        )
-
-
-def get_card_back(name: str = "card_back", width: int = 80) -> Optional[pygame.Surface]:
-    if name not in _BASE_IMAGES:
-        return None
-    key = (name, width)
-    if key not in _CARD_CACHE:
-        base = _BASE_IMAGES[name]
-        ratio = width / base.get_width()
-        _CARD_CACHE[key] = pygame.transform.smoothscale(
-            base, (width, int(base.get_height() * ratio))
-        )
-    return _CARD_CACHE[key]
-
-
-def get_card_image(card: Card, width: int) -> pygame.Surface:
-    key = _image_key(card)
-    if (key, width) not in _CARD_CACHE:
-        base = _BASE_IMAGES.get(key)
-        if base is None:
-            return None
-        ratio = width / base.get_width()
-        _CARD_CACHE[(key, width)] = pygame.transform.smoothscale(
-            base, (width, int(base.get_height() * ratio))
-        )
-    return _CARD_CACHE[(key, width)]
-
-
-# ---------------------------------------------------------------------------
-# Sprite classes
-# ---------------------------------------------------------------------------
-
-
-class CardSprite(pygame.sprite.Sprite):
-    def __init__(self, card: Card, pos: Tuple[int, int], width: int = 80) -> None:
-        super().__init__()
-        img = get_card_image(card, width)
-        if img is None:
-            # Render a text fallback
-            font = pygame.font.SysFont(None, 20)
-            img = font.render(str(card), True, (0, 0, 0), (255, 255, 255))
-
-        border = 2
-        w, h = img.get_size()
-        surf = pygame.Surface((w + border * 2, h + border * 2), pygame.SRCALPHA)
-        rect = surf.get_rect()
-        pygame.draw.rect(surf, (220, 220, 220), rect, border_radius=5)
-        pygame.draw.rect(surf, (0, 0, 0), rect, width=1, border_radius=5)
-        surf.blit(img, (border, border))
-
-        self.image = surf
-        self.rect = self.image.get_rect(topleft=pos)
-        self.card = card
-        self.selected = False
-
-    def toggle(self) -> None:
-        self.selected = not self.selected
-        offset = -10 if self.selected else 10
-        self.rect.move_ip(0, offset)
-
-    def draw_shadow(
-        self,
-        surface: pygame.Surface,
-        offset: Tuple[int, int] = (5, 5),
-        blur: int = 2,
-        alpha: int = 80,
-    ) -> None:
-        """Draw a simple blurred shadow beneath the card."""
-        shadow = pygame.Surface(self.image.get_size(), pygame.SRCALPHA)
-        shadow.fill((0, 0, 0))
-        shadow.set_alpha(alpha)
-        for dx in range(-blur, blur + 1):
-            for dy in range(-blur, blur + 1):
-                rect = self.rect.move(offset[0] + dx, offset[1] + dy)
-                surface.blit(shadow, rect)
-
-
-def draw_glow(
-    surface: pygame.Surface,
-    rect: pygame.Rect,
-    color: Tuple[int, int, int],
-    radius: int = 8,
-    alpha: int = 100,
-) -> None:
-    """Draw a subtle glow effect behind a rect."""
-    glow = pygame.Surface(
-        (rect.width + radius * 2, rect.height + radius * 2), pygame.SRCALPHA
-    )
-    for dx in range(-radius, radius + 1):
-        for dy in range(-radius, radius + 1):
-            dist = dx * dx + dy * dy
-            if dist <= radius * radius:
-                glow_x = rect.x - radius + dx
-                glow_y = rect.y - radius + dy
-                glow.fill((*color, alpha // (1 + dist)), special_flags=pygame.BLEND_RGBA_ADD)
-                surface.blit(glow, (glow_x, glow_y))
-
-
-class CardBackSprite(pygame.sprite.Sprite):
-    def __init__(
-        self,
-        pos: Tuple[int, int],
-        width: int = 80,
-        name: str = "card_back",
-        rotation: int = 0,
-    ) -> None:
-        super().__init__()
-        img = get_card_back(name, width)
-        if img is None:
-            font = pygame.font.SysFont(None, 20)
-            img = font.render("[]", True, (0, 0, 0), (255, 255, 255))
-        if rotation:
-            img = pygame.transform.rotate(img, rotation)
-        self.image = img
-        self.rect = self.image.get_rect(topleft=pos)
-
-
-# ---------------------------------------------------------------------------
-# Simple button and overlay helpers
-# ---------------------------------------------------------------------------
-
-
-class Button:
-    """Basic rectangular button used by overlays."""
-
-    def __init__(
-        self,
-        text: str,
-        rect: pygame.Rect,
-        callback: Callable[[], None],
-        font: pygame.font.Font,
-        enabled: bool = True,
-    ) -> None:
-        self.text = text
-        self.rect = rect
-        self.callback = callback
-        self.font = font
-        self.enabled = enabled
-        self.hovered = False
-        self.selected = False
-
-    def draw(self, surface: pygame.Surface) -> None:
-        if not self.enabled:
-            color = (150, 150, 150)
-            text_color = (100, 100, 100)
-        elif self.selected:
-            color = (255, 220, 120)
-            text_color = (0, 0, 0)
-        elif self.hovered:
-            color = (220, 220, 220)
-            text_color = (0, 0, 0)
-        else:
-            color = (200, 200, 200)
-            text_color = (0, 0, 0)
-        pygame.draw.rect(surface, color, self.rect)
-        pygame.draw.rect(surface, (0, 0, 0), self.rect, 2)
-        txt = self.font.render(self.text, True, text_color)
-        surface.blit(txt, txt.get_rect(center=self.rect.center))
-
-    def handle_event(self, event: pygame.event.Event) -> None:
-        if event.type == pygame.MOUSEMOTION:
-            self.hovered = self.rect.collidepoint(event.pos)
-        if (
-            self.enabled
-            and event.type == pygame.MOUSEBUTTONDOWN
-            and self.rect.collidepoint(event.pos)
-        ):
-            self.callback()
-
-
-class Overlay:
-    """Base overlay class managing a list of buttons."""
-
-    def __init__(
-        self, view: "GameView", back_cb: Optional[Callable[[], None]] = None
-    ) -> None:
-        self.view = view
-        self.buttons: List[Button] = []
-        self.focus_idx = 0
-        self.back_callback = back_cb
-
-    def resize(self) -> None:
-        """Recalculate layout based on the current screen size."""
-        pass
-
-    def draw(self, surface: pygame.Surface) -> None:
-        for idx, btn in enumerate(self.buttons):
-            btn.selected = idx == self.focus_idx
-            btn.draw(surface)
-
-    def back(self) -> None:
-        if self.back_callback:
-            self.back_callback()
-
-    def handle_event(self, event: pygame.event.Event) -> None:
-        if event.type == pygame.MOUSEMOTION:
-            for i, btn in enumerate(self.buttons):
-                if btn.rect.collidepoint(event.pos):
-                    self.focus_idx = i
-        for btn in self.buttons:
-            btn.handle_event(event)
-
-        if event.type == pygame.KEYDOWN:
-            if event.key in (pygame.K_DOWN, pygame.K_s):
-                self.focus_idx = (self.focus_idx + 1) % len(self.buttons)
-            elif event.key in (pygame.K_UP, pygame.K_w):
-                self.focus_idx = (self.focus_idx - 1) % len(self.buttons)
-            elif event.key in (pygame.K_RETURN, pygame.K_KP_ENTER):
-                btn = self.buttons[self.focus_idx]
-                if btn.enabled:
-                    btn.callback()
-            elif event.key == pygame.K_ESCAPE:
-                self.back()
-        elif event.type == pygame.JOYHATMOTION:
-            if event.value[1] > 0:
-                self.focus_idx = (self.focus_idx - 1) % len(self.buttons)
-            elif event.value[1] < 0:
-                self.focus_idx = (self.focus_idx + 1) % len(self.buttons)
-        elif event.type == pygame.JOYBUTTONDOWN:
-            if event.button == 0:
-                btn = self.buttons[self.focus_idx]
-                if btn.enabled:
-                    btn.callback()
-            elif event.button == 1:
-                self.back()
-
-
-class MainMenuOverlay(Overlay):
-    """Initial game menu."""
-
-    def __init__(self, view: "GameView") -> None:
-        super().__init__(view, view.close_overlay)
-        self._layout()
-
-    def resize(self) -> None:
-        self._layout()
-
-    def _layout(self) -> None:
-        w, h = self.view.screen.get_size()
-        font = self.view.font
-        bx = w // 2 - 100
-        by = h // 2 - 120
-        self.buttons = [
-            Button(
-                "New Game", pygame.Rect(bx, by, 200, 40), self.view.restart_game, font
-            ),
-            Button(
-                "Load Game",
-                pygame.Rect(bx, by + 50, 200, 40),
-                self.view.load_game,
-                font,
-            ),
-            Button(
-                "Settings",
-                pygame.Rect(bx, by + 100, 200, 40),
-                self.view.show_settings,
-                font,
-            ),
-            Button(
-                "How to Play",
-                pygame.Rect(bx, by + 150, 200, 40),
-                lambda: self.view.show_how_to_play(from_menu=True),
-                font,
-            ),
-            Button(
-                "Quit", pygame.Rect(bx, by + 200, 200, 40), self.view.quit_game, font
-            ),
-        ]
-        if self.focus_idx >= len(self.buttons):
-            self.focus_idx = max(0, len(self.buttons) - 1)
-
-
-class InGameMenuOverlay(Overlay):
-    """Menu accessible during gameplay via the Settings button."""
-
-    def __init__(self, view: "GameView") -> None:
-        super().__init__(view, view.close_overlay)
-        self._layout()
-
-    def resize(self) -> None:
-        self._layout()
-
-    def _layout(self) -> None:
-        w, h = self.view.screen.get_size()
-        font = self.view.font
-        bx = w // 2 - 100
-        by = h // 2 - 150
-        self.buttons = [
-            Button(
-                "Resume Game",
-                pygame.Rect(bx, by, 200, 40),
-                self.view.close_overlay,
-                font,
-            ),
-            Button(
-                "Save Game",
-                pygame.Rect(bx, by + 50, 200, 40),
-                self.view.save_game,
-                font,
-            ),
-            Button(
-                "Load Game",
-                pygame.Rect(bx, by + 100, 200, 40),
-                self.view.load_game,
-                font,
-            ),
-            Button(
-                "Game Settings",
-                pygame.Rect(bx, by + 150, 200, 40),
-                self.view.show_settings,
-                font,
-            ),
-            Button(
-                "Return to Main Menu",
-                pygame.Rect(bx, by + 200, 200, 40),
-                self.view.confirm_return_to_menu,
-                font,
-            ),
-            Button(
-                "Quit Game",
-                pygame.Rect(bx, by + 250, 200, 40),
-                self.view.confirm_quit,
-                font,
-            ),
-        ]
-        if self.focus_idx >= len(self.buttons):
-            self.focus_idx = max(0, len(self.buttons) - 1)
-
-
-class SettingsOverlay(Overlay):
-    """Top level settings menu."""
-
-    def __init__(self, view: "GameView") -> None:
-        super().__init__(view, view.close_overlay)
-        self._layout()
-
-    def resize(self) -> None:
-        self._layout()
-
-    def _layout(self) -> None:
-        w, h = self.view.screen.get_size()
-        font = self.view.font
-        bx = w // 2 - 120
-        by = h // 2 - 120
-        self.buttons = [
-            Button(
-                "Game Settings",
-                pygame.Rect(bx, by, 240, 40),
-                self.view.show_game_settings,
-                font,
-            ),
-            Button(
-                "Graphics",
-                pygame.Rect(bx, by + 50, 240, 40),
-                self.view.show_graphics,
-                font,
-            ),
-            Button(
-                "Audio", pygame.Rect(bx, by + 100, 240, 40), self.view.show_audio, font
-            ),
-            Button(
-                "Game Tutorial",
-                pygame.Rect(bx, by + 150, 240, 40),
-                lambda: self.view.show_tutorial(from_menu=False),
-                font,
-            ),
-            Button(
-                "Back",
-                pygame.Rect(bx, by + 200, 240, 40),
-                self.view.close_overlay,
-                font,
-            ),
-        ]
-
-
-class GameSettingsOverlay(Overlay):
-    def __init__(self, view: "GameView") -> None:
-        super().__init__(view, view.show_settings)
-        self._layout()
-
-    def resize(self) -> None:
-        self._layout()
-
-    def _layout(self) -> None:
-        w, h = self.view.screen.get_size()
-        font = self.view.font
-        bx = w // 2 - 120
-        by = h // 2 - 220
-
-        def cycle(attr: str, options: List, label: str) -> Callable[[], None]:
-            def callback(b: Button) -> Callable[[], None]:
-                def inner() -> None:
-                    cur = getattr(self.view, attr)
-                    idx = options.index(cur)
-                    cur = options[(idx + 1) % len(options)]
-                    setattr(self.view, attr, cur)
-                    self.view.apply_options()
-                    b.text = f"{label}: {cur if not isinstance(cur, bool) else ('On' if cur else 'Off')}"
-
-                return inner
-
-            return callback
-
-        self.buttons = []
-
-        def make_button(offset: int, attr: str, opts: List, label: str) -> None:
-            text = getattr(self.view, attr)
-            if isinstance(text, bool):
-                text = "On" if text else "Off"
-            btn = Button(
-                f"{label}: {text}",
-                pygame.Rect(bx, by + offset, 240, 40),
-                lambda: None,
-                font,
-            )
-            btn.callback = cycle(attr, opts, label)(btn)
-            self.buttons.append(btn)
-
-        make_button(0, "ai_level", ["Easy", "Normal", "Hard", "Expert"], "AI Level")
-        make_button(
-            50,
-            "ai_personality",
-            ["balanced", "aggressive", "defensive", "random"],
-            "Personality",
-        )
-        make_button(100, "ai_lookahead", [False, True], "Lookahead")
-        make_button(150, "animation_speed", [0.5, 1.0, 2.0], "Anim Speed")
-        make_button(200, "sort_mode", ["rank", "suit"], "Sort Mode")
-        self.buttons.append(
-            Button(
-                "House Rules",
-                pygame.Rect(bx, by + 250, 240, 40),
-                lambda: self.view.show_rules(from_menu=False),
-                font,
-            )
-        )
-        btn = Button(
-            "Back", pygame.Rect(bx, by + 300, 240, 40), self.view.show_settings, font
-        )
-        self.buttons.append(btn)
-
-
-class GraphicsOverlay(Overlay):
-    def __init__(self, view: "GameView") -> None:
-        super().__init__(view, view.show_settings)
-        self._layout()
-
-    def resize(self) -> None:
-        self._layout()
-
-    def _layout(self) -> None:
-        w, h = self.view.screen.get_size()
-        font = self.view.font
-        bx = w // 2 - 120
-        by = h // 2 - 120
-
-        def cycle(attr: str, options: List, label: str) -> Callable[[], None]:
-            def callback(btn: Button) -> Callable[[], None]:
-                def inner() -> None:
-                    cur = getattr(self.view, attr)
-                    idx = options.index(cur)
-                    cur = options[(idx + 1) % len(options)]
-                    setattr(self.view, attr, cur)
-                    self.view.apply_options()
-                    if isinstance(cur, bool):
-                        cur_txt = "On" if cur else "Off"
-                    else:
-                        cur_txt = cur
-                    btn.text = f"{label}: {cur_txt}"
-
-                return inner
-
-            return callback
-
-        make_color = list(TABLE_THEMES.keys())
-        make_card_color = list_card_back_colors() or ["blue"]
-        make_table_tex = list_table_textures() or ["table_img"]
-
-        self.buttons = []
-
-        def make_button(offset: int, attr: str, opts: List, label: str) -> None:
-            text = getattr(self.view, attr)
-            btn = Button(
-                f"{label}: {text}",
-                pygame.Rect(bx, by + offset, 240, 40),
-                lambda: None,
-                font,
-            )
-            btn.callback = cycle(attr, opts, label)(btn)
-            self.buttons.append(btn)
-
-        make_button(0, "table_color_name", make_color, "Table Color")
-        make_button(50, "card_color", make_card_color, "Card Back")
-        make_button(100, "table_texture_name", make_table_tex, "Table Tex")
-        make_button(150, "colorblind_mode", [False, True], "Colorblind")
-        make_button(200, "fullscreen", [False, True], "Fullscreen")
-        btn = Button(
-            "Back", pygame.Rect(bx, by + 250, 240, 40), self.view.show_settings, font
-        )
-        self.buttons.append(btn)
-
-        self.table_preview_rect = pygame.Rect(bx + 250, by, 60, 40)
-        self.back_preview_rect = pygame.Rect(bx + 250, by + 50, 40, 60)
-
-    def draw(self, surface: pygame.Surface) -> None:
-        super().draw(surface)
-        if self.view.table_image:
-            tex = pygame.transform.smoothscale(
-                self.view.table_image, self.table_preview_rect.size
-            )
-            surface.blit(tex, self.table_preview_rect)
-        else:
-            pygame.draw.rect(surface, self.view.table_color, self.table_preview_rect)
-
-        back = get_card_back(self.view.card_back_name, self.back_preview_rect.width)
-        if back:
-            img = pygame.transform.smoothscale(back, self.back_preview_rect.size)
-            surface.blit(img, self.back_preview_rect)
-
-
-class AudioOverlay(Overlay):
-    def __init__(self, view: "GameView") -> None:
-        super().__init__(view, view.show_settings)
-        self._layout()
-
-    def resize(self) -> None:
-        self._layout()
-
-    def _layout(self) -> None:
-        w, h = self.view.screen.get_size()
-        font = self.view.font
-        bx = w // 2 - 120
-        by = h // 2 - 120
-
-        def cycle(attr: str, options: List, label: str) -> Callable[[], None]:
-            def callback(btn: Button) -> Callable[[], None]:
-                def inner() -> None:
-                    cur = getattr(self.view, attr)
-                    idx = options.index(cur)
-                    cur = options[(idx + 1) % len(options)]
-                    setattr(self.view, attr, cur)
-                    self.view.apply_options()
-                    btn.text = f"{label}: {cur if not isinstance(cur, bool) else ('On' if cur else 'Off')}"
-
-                return inner
-
-            return callback
-
-        self.buttons = []
-
-        def make_button(offset: int, attr: str, opts: List, label: str) -> None:
-            text = getattr(self.view, attr)
-            if isinstance(text, bool):
-                text = "On" if text else "Off"
-            btn = Button(
-                f"{label}: {text}",
-                pygame.Rect(bx, by + offset, 240, 40),
-                lambda: None,
-                font,
-            )
-            btn.callback = cycle(attr, opts, label)(btn)
-            self.buttons.append(btn)
-
-        make_button(0, "sound_enabled", [True, False], "Sound")
-        make_button(50, "music_enabled", [True, False], "Music")
-        make_button(100, "fx_volume", [0.5, 0.75, 1.0], "FX Vol")
-        make_button(150, "music_volume", [0.5, 0.75, 1.0], "Music Vol")
-        make_button(200, "music_track", list_music_tracks() or [self.view.music_track], "Track")
-        btn = Button(
-            "Back", pygame.Rect(bx, by + 250, 240, 40), self.view.show_settings, font
-        )
-        self.buttons.append(btn)
-
-
-class RulesOverlay(Overlay):
-    """Overlay providing toggles for optional house rules."""
-
-    def __init__(self, view: "GameView", back_cb: Callable[[], None]) -> None:
-        super().__init__(view, back_cb)
-        self._layout()
-
-    def resize(self) -> None:
-        self._layout()
-
-    def _layout(self) -> None:
-        w, h = self.view.screen.get_size()
-        font = self.view.font
-        bx = w // 2 - 120
-        by = h // 2 - 150
-
-        def toggle(attr: str, label: str) -> Callable[[Button], Callable[[], None]]:
-            def cb(btn: Button) -> Callable[[], None]:
-                def inner() -> None:
-                    cur = getattr(self.view, attr)
-                    setattr(self.view, attr, not cur)
-                    self.view.apply_options()
-                    btn.text = f"{label}: {'On' if getattr(self.view, attr) else 'Off'}"
-
-                return inner
-
-            return cb
-
-        self.buttons = []
-
-        def make_button(offset: int, attr: str, label: str) -> None:
-            state = getattr(self.view, attr)
-            text = f"{label}: {'On' if state else 'Off'}"
-            btn = Button(
-                text, pygame.Rect(bx, by + offset, 240, 40), lambda: None, font
-            )
-            btn.callback = toggle(attr, label)(btn)
-            self.buttons.append(btn)
-
-        make_button(0, "rule_flip_suit_rank", "Flip Suit Rank")
-        make_button(50, "rule_no_2s", "No 2s in straights")
-        self.buttons.append(
-            Button("Back", pygame.Rect(bx, by + 100, 240, 40), self.back_callback, font)
-        )
-
-
-class HowToPlayOverlay(Overlay):
-    """Display a short summary of the basic rules."""
-
-    def __init__(self, view: "GameView", back_cb: Callable[[], None]) -> None:
-        super().__init__(view, back_cb)
-        self._layout()
-
-    def resize(self) -> None:
-        self._layout()
-
-    def _layout(self) -> None:
-        w, h = self.view.screen.get_size()
-        font = self.view.font
-        bx = w // 2 - 100
-        by = h // 2 + 60
-        self.buttons = [
-            Button("Back", pygame.Rect(bx, by, 200, 40), self.back_callback, font)
-        ]
-
-    def draw(self, surface: pygame.Surface) -> None:
-        super().draw(surface)
-        w, h = surface.get_size()
-        font = pygame.font.SysFont(None, 20)
-        lines = [
-            "Each player starts with 13 cards.",
-            "Play higher combinations to beat opponents.",
-            "First to shed all cards wins the round.",
-        ]
-        y = h // 2 - 40
-        for line in lines:
-            img = font.render(line, True, (255, 255, 255))
-            surface.blit(img, img.get_rect(center=(w // 2, y)))
-            y += 24
-
-
-class TutorialOverlay(Overlay):
-    """Explain gameplay via numbered steps."""
-
-    def __init__(self, view: "GameView", back_cb: Callable[[], None]) -> None:
-        super().__init__(view, back_cb)
-        self._layout()
-
-    def resize(self) -> None:
-        self._layout()
-
-    def _layout(self) -> None:
-        w, h = self.view.screen.get_size()
-        font = self.view.font
-        bx = w // 2 - 100
-        by = h // 2 + 60
-        self.buttons = [
-            Button("Back", pygame.Rect(bx, by, 200, 40), self.back_callback, font)
-        ]
-
-    def draw(self, surface: pygame.Surface) -> None:
-        super().draw(surface)
-        w, h = surface.get_size()
-        font = pygame.font.SysFont(None, 20)
-        steps = [
-            "1. Select cards with mouse or arrow keys.",
-            "2. Press Play to submit your move.",
-            "3. Beat opponents until you run out of cards.",
-        ]
-        y = h // 2 - 40
-        for line in steps:
-            img = font.render(line, True, (255, 255, 255))
-            surface.blit(img, img.get_rect(center=(w // 2, y)))
-            y += 24
-
-
-class SavePromptOverlay(Overlay):
-    """Prompt the user to save before performing an action."""
-
-    def __init__(
-        self, view: "GameView", action: Callable[[], None], label: str
-    ) -> None:
-        super().__init__(view, view.close_overlay)
-        self.action = action
-        self.label = label
-        self._layout()
-
-    def resize(self) -> None:
-        self._layout()
-
-    def _layout(self) -> None:
-        w, h = self.view.screen.get_size()
-        font = self.view.font
-        bx = w // 2 - 120
-        by = h // 2
-        self.buttons = [
-            Button(
-                f"Save and {self.label}",
-                pygame.Rect(bx, by, 240, 40),
-                self._save_then_action,
-                font,
-            ),
-            Button(
-                f"{self.label} Without Saving",
-                pygame.Rect(bx, by + 50, 240, 40),
-                self.action,
-                font,
-            ),
-            Button(
-                "Cancel",
-                pygame.Rect(bx, by + 100, 240, 40),
-                self.view.close_overlay,
-                font,
-            ),
-        ]
-
-    def _save_then_action(self) -> None:
-        self.view.save_game()
-        self.action()
-
-    def draw(self, surface: pygame.Surface) -> None:
-        w, h = surface.get_size()
-        font = pygame.font.SysFont(None, 24)
-        msg = f"Save your game before {self.label.lower()}?"
-        img = font.render(msg, True, (255, 255, 255))
-        surface.blit(img, img.get_rect(center=(w // 2, h // 2 - 60)))
-        super().draw(surface)
-
-
-class ProfileOverlay(Overlay):
-    """Select or create a player profile."""
-
-    def __init__(self, view: "GameView") -> None:
-        super().__init__(view, None)
-        self._layout()
-
-    def resize(self) -> None:
-        self._layout()
-
-    def _layout(self) -> None:
-        w, h = self.view.screen.get_size()
-        font = self.view.font
-        names = list(self.view.win_counts.keys())
-        bx = w // 2 - 100
-        start_y = h // 2 - (len(names) + 1) * 25
-        self.buttons = [
-            Button(
-                name,
-                pygame.Rect(bx, start_y + i * 50, 200, 40),
-                lambda n=name: self.select(n),
-                font,
-            )
-            for i, name in enumerate(names)
-        ]
-        self.buttons.append(
-            Button(
-                "New Profile",
-                pygame.Rect(bx, start_y + len(names) * 50, 200, 40),
-                self.new_profile,
-                font,
-            )
-        )
-
-    def select(self, name: str) -> None:
-        self.view.player_name = name
-        self.view.apply_options()
-        self.view._save_options()
-        self.view.show_menu()
-
-    def new_profile(self) -> None:
-        base = "Player"
-        idx = 1
-        names = set(self.view.win_counts)
-        while f"{base}{idx}" in names:
-            idx += 1
-        name = f"{base}{idx}"
-        self.view.win_counts[name] = 0
-        self.select(name)
-
-
-class GameOverOverlay(Overlay):
-    def __init__(self, view: "GameView", winner: str) -> None:
-        super().__init__(view, None)
-        self.winner = winner
-        self.rankings = view.game.get_rankings()
-        self._layout()
-
-    def resize(self) -> None:
-        self._layout()
-
-    def _layout(self) -> None:
-        w, h = self.view.screen.get_size()
-        font = self.view.font
-        bx = w // 2 - 100
-        by = h // 2 + 40
-        self.buttons = [
-            Button(
-                "Play Again", pygame.Rect(bx, by, 200, 40), self.view.restart_game, font
-            ),
-            Button(
-                "Quit", pygame.Rect(bx, by + 50, 200, 40), self.view.quit_game, font
-            ),
-        ]
-
-    def draw(self, surface: pygame.Surface) -> None:
-        w, h = surface.get_size()
-        font = pygame.font.SysFont(None, 32)
-        txt = font.render(f"{self.winner} wins!", True, (255, 255, 255))
-        surface.blit(txt, txt.get_rect(center=(w // 2, h // 2 - 60)))
-        rank_lines = [f"{i+1}. {n} ({c})" for i, (n, c) in enumerate(self.rankings)]
-        y = h // 2 - 20
-        for line in rank_lines:
-            img = font.render(line, True, (255, 255, 255))
-            surface.blit(img, img.get_rect(center=(w // 2, y)))
-            y += 30
-        super().draw(surface)
-
-
-# ---------------------------------------------------------------------------
-# Main game view
-# ---------------------------------------------------------------------------
-
-
-class GameView:
+class GameView(AnimationMixin):
     TABLE_COLOR = TABLE_THEMES["darkgreen"]
+
 
     def __init__(self, width: int = 1024, height: int = 768) -> None:
         pygame.init()
@@ -1160,190 +206,6 @@ class GameView:
             self.overlay.draw(self.screen)
         self.draw_score_overlay()
         pygame.display.flip()
-
-    def _animate_sprites(
-        self, sprites: List[CardSprite], dest: Tuple[int, int], frames: int = 15
-    ) -> None:
-        """Move ``sprites`` toward ``dest`` over ``frames`` steps."""
-        if not sprites:
-            return
-        frames = max(1, int(frames / self.animation_speed))
-        starts = [sp.rect.center for sp in sprites]
-        for i in range(frames):
-            t = (i + 1) / frames
-            for sp, (sx, sy) in zip(sprites, starts):
-                sp.rect.center = (
-                    int(sx + (dest[0] - sx) * t),
-                    int(sy + (dest[1] - sy) * t),
-                )
-            self._draw_frame()
-            pygame.event.pump()
-            self.clock.tick(60)
-
-    def _animate_bounce(
-        self, sprites: List[CardSprite], scale: float = 1.2, frames: int = 6
-    ) -> None:
-        """Briefly scale ``sprites`` up then back down for a bounce effect."""
-        if not sprites:
-            return
-        frames = max(1, int(frames / self.animation_speed))
-        half = frames // 2 or 1
-        originals = [(sp.image, sp.rect.copy()) for sp in sprites]
-        for i in range(frames):
-            # progress goes 0->1 then 1->0
-            t = (i + 1) / half if i < half else (frames - i) / half
-            factor = 1 + (scale - 1) * t
-            for sp, (img, rect) in zip(sprites, originals):
-                w, h = rect.size
-                if isinstance(img, pygame.Surface):
-                    scaled = pygame.transform.smoothscale(
-                        img, (int(w * factor), int(h * factor))
-                    )
-                else:
-                    scaled = img
-                sp.image = scaled
-                sp.rect = scaled.get_rect(center=rect.center)
-            self._draw_frame()
-            pygame.event.pump()
-            self.clock.tick(60)
-        # restore originals
-        for sp, (img, rect) in zip(sprites, originals):
-            sp.image = img
-            sp.rect = rect
-
-    def _animate_back(
-        self, start: Tuple[int, int], dest: Tuple[int, int], frames: int = 15
-    ) -> None:
-        """Animate a card back image from ``start`` to ``dest``."""
-        img = get_card_back(self.card_back_name)
-        if img is None:
-            return
-        frames = max(1, int(frames / self.animation_speed))
-        rect = img.get_rect(center=start)
-        for i in range(frames):
-            t = (i + 1) / frames
-            rect.center = (
-                int(start[0] + (dest[0] - start[0]) * t),
-                int(start[1] + (dest[1] - start[1]) * t),
-            )
-            self._draw_frame()
-            self.screen.blit(img, rect)
-            pygame.display.flip()
-            pygame.event.pump()
-            self.clock.tick(60)
-        dummy = types.SimpleNamespace(image=img, rect=rect)
-        self._animate_bounce([dummy])
-
-    def _animate_flip(
-        self, sprites: List[CardSprite], dest: Tuple[int, int], frames: int = 15
-    ) -> None:
-        """Move ``sprites`` to ``dest`` while flipping from back to front."""
-        if not sprites:
-            return
-        frames = max(1, int(frames / self.animation_speed))
-        starts = [sp.rect.center for sp in sprites]
-        fronts = [sp.image for sp in sprites]
-        back = get_card_back(self.card_back_name, sprites[0].rect.width)
-        for i in range(frames):
-            t = (i + 1) / frames
-            for sp, (sx, sy) in zip(sprites, starts):
-                sp.rect.center = (
-                    int(sx + (dest[0] - sx) * t),
-                    int(sy + (dest[1] - sy) * t),
-                )
-            self._draw_frame()
-            for sp, front in zip(sprites, fronts):
-                img = back if back is not None and t < 0.5 else front
-                rect = img.get_rect(center=sp.rect.center)
-                self.screen.blit(img, rect)
-            pygame.display.flip()
-            pygame.event.pump()
-            self.clock.tick(60)
-        self._animate_bounce(sprites)
-
-    def _animate_select(self, sprite: CardSprite, up: bool) -> None:
-        offset = -10 if up else 10
-        dest = (sprite.rect.centerx, sprite.rect.centery + offset)
-        self._animate_sprites([sprite], dest, frames=5)
-
-    def _highlight_turn(self, idx: int, frames: int = 10) -> None:
-        """Flash the active player's name for visual emphasis."""
-        x, y = self._player_pos(idx)
-        rect = pygame.Rect(0, 0, 140, 30)
-        if idx == 0:
-            rect.midbottom = (x, y)
-        elif idx == 1:
-            rect.midtop = (x, y)
-        elif idx == 2:
-            rect.midleft = (x, y)
-        else:
-            rect.midright = (x, y)
-        frames = max(1, int(frames / self.animation_speed))
-        for i in range(frames):
-            self._draw_frame()
-            overlay = pygame.Surface(rect.size, pygame.SRCALPHA)
-            alpha = max(0, 200 - i * 20)
-            center = overlay.get_rect().center
-            radius = 11 + int(3 * math.sin(math.pi * i / frames))
-            if hasattr(overlay, "get_width"):
-                pygame.draw.circle(
-                    overlay, (255, 255, 0, alpha), center, radius, width=3
-                )
-            self.screen.blit(overlay, rect.topleft)
-            pygame.display.flip()
-            pygame.event.pump()
-            self.clock.tick(60)
-
-    def _transition_overlay(
-        self,
-        old: Optional[Overlay],
-        new: Overlay,
-        frames: int = 20,
-        slide: bool = False,
-    ) -> None:
-        """Animate transition between two overlays."""
-        if old is None:
-            return
-        frames = max(1, int(frames / self.animation_speed))
-        w, h = self.screen.get_size()
-
-        def render(ov: Overlay) -> pygame.Surface:
-            surf = pygame.Surface((w, h), pygame.SRCALPHA)
-            bg = pygame.Surface((w, h), pygame.SRCALPHA)
-            bg.fill((0, 0, 0, 180))
-            surf.blit(bg, (0, 0))
-            ov.draw(surf)
-            return surf
-
-        from_surf = render(old)
-        to_surf = render(new)
-
-        current = self.overlay
-        # Draw the base screen once and reuse it during the animation
-        self.overlay = None
-        self._draw_frame()
-        base = self.screen.copy()
-        self.overlay = current
-        for i in range(frames):
-            progress = (i + 1) / frames
-            self.screen.blit(base, (0, 0))
-            if slide:
-                offset = int(w * (1 - progress))
-                self.screen.blit(from_surf, (-offset, 0))
-                self.screen.blit(to_surf, (w - offset, 0))
-            else:
-                fs = from_surf.copy()
-                fs.set_alpha(int(255 * (1 - progress)))
-                ts = to_surf.copy()
-                ts.set_alpha(int(255 * progress))
-                self.screen.blit(fs, (0, 0))
-                self.screen.blit(ts, (0, 0))
-            pygame.display.flip()
-            pygame.event.pump()
-            self.clock.tick(60)
-
-        self.overlay = new
-        self._draw_frame()
 
     # Layout helpers --------------------------------------------------
     def _layout_zones(self) -> None:
